Move password reset routes to auth, simplify naming
