+++ server/routes/api/v1/auth/login.js
@@ -83,10 +83,4 @@ export default async function (fastify, _opts) {
       reply.send(data);
     },
   );
-
-  // add a logout route
-  fastify.get('/logout', (request, reply) => {
-    request.session.delete();
-    reply.send();
-  });
 }
