+++ server/routes/api/v1/auth/logout.js
@@ -1,0 +1,7 @@
+export default async function (fastify, _opts) {
+  // add a logout route
+  fastify.get('/logout', (request, reply) => {
+    request.session.delete();
+    reply.send();
+  });
+}
