+++ server/routes/api/v1/auth/password/get.js
@@ -1,9 +1,9 @@
 import { StatusCodes } from 'http-status-codes';
-import User from '../../../../models/user.js';
+import User from '../../../../../models/user.js';
 
 export default async function (fastify, _opts) {
   fastify.get(
-    '/verify-password-reset/:token',
+    '/:token',
     {
       schema: {
         params: {
